requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-allauth==0.48.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-phonenumber-field==6.1.0
django-suit==2.0a1
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.4.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.4.1
psutil==5.9.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.7
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
Werkzeug==2.0.3
wrapt==1.10.11
xlwt==1.3.0
yarg==0.1.9
